requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
joblib
opencv-python
scikit-learn
streamlit
python-multipart
locust
